requirements/prod.txt
# Everything needed in production

# Flask
flask==1.1.0
flask-cors
flask_restplus
flask-jwt-extended
werkzeug==0.16.1
click>=5.0

dateutils == 0.6.6
# Database
Flask-SQLAlchemy==2.3.2
SQLAlchemy==1.3.13
psycopg2==2.8.3
Flask-SQLAlchemy-Caching==1.0.4
# Migrations
Flask-Migrate==2.2.1
sqlalchemy-mixins==1.2.1
dj_database_url

# aio
gevent==1.4.0

# log
nb-log==4.1

# Auth
Flask-Login==0.4.1
Flask-Bcrypt==0.7.1
passlib

# Caching
Flask-Caching>=1.0.0

# Environment variable parsing
python-dotenv

# numeric
scipy

# APIs
marshmallow==3.2.0
flask-marshmallow==0.10.1
marshmallow-sqlalchemy==0.19.0
# model
pandas==0.24.2
Pillow==7.1.2
# config
pyyaml==5.1.1
DictObject==1.1.1
# task engine
tox
celery[redis]

# aipspec
apispec[yaml]
apispec-webframeworks
flasgger==0.9.4
#excel
openpyxl==2.6.4
xlrd
# http
requests
tqdm

# reg1exp
regex

# utils
func-timeout==4.3.5
retry==0.9.2

#nlp
jieba==0.42.1
py_stringmatching

# spider
python-bidi

# record deduplication
dedupe